# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EvoAgi",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/teams/{team_id}/members/": {
      "get": {
        "tags": [
          "members"
        ],
        "summary": "Read Members",
        "description": "Retrieve members from team.",
        "operationId": "read_members",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembersOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "members"
        ],
        "summary": "Create Member",
        "description": "Create new member.",
        "operationId": "create_member",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{team_id}/members/{id}": {
      "get": {
        "tags": [
          "members"
        ],
        "summary": "Read Member",
        "description": "Get member by ID.",
        "operationId": "read_member",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "members"
        ],
        "summary": "Update Member",
        "description": "Update a member.",
        "operationId": "update_member",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "members"
        ],
        "summary": "Delete Member",
        "description": "Delete a member.",
        "operationId": "delete_member",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/provider/": {
      "post": {
        "tags": [
          "provider"
        ],
        "summary": "Create Provider",
        "operationId": "create_provider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProvider"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provider/{model_provider_id}": {
      "get": {
        "tags": [
          "provider"
        ],
        "summary": "Read Provider",
        "operationId": "read_provider",
        "parameters": [
          {
            "name": "model_provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProvider"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "provider"
        ],
        "summary": "Update Provider",
        "operationId": "update_provider",
        "parameters": [
          {
            "name": "model_provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProviderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProvider"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "provider"
        ],
        "summary": "Delete Provider",
        "operationId": "delete_provider",
        "parameters": [
          {
            "name": "model_provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProvider"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provider/withmodels/{model_provider_id}": {
      "get": {
        "tags": [
          "provider"
        ],
        "summary": "Read Provider With Model List",
        "operationId": "read_provider_with_model_list",
        "parameters": [
          {
            "name": "model_provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProviderWithModelsListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberCreate": {
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,64}$",
            "title": "Name"
          },
          "backstory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backstory"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "owner_of": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Of"
          },
          "position_x": {
            "type": "number",
            "title": "Position X"
          },
          "position_y": {
            "type": "number",
            "title": "Position Y"
          },
          "source": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "zhipuai"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "glm-4-0520"
          },
          "openai_api_key": {
            "type": "string",
            "title": "Openai Api Key",
            "default": "fe1f2097b7284bd4baa1284be8d54aea.6VOvX4efbye8M6m0"
          },
          "openai_api_base": {
            "type": "string",
            "title": "Openai Api Base",
            "default": "https://open.bigmodel.cn/api/paas/v4"
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "default": 0.1
          },
          "interrupt": {
            "type": "boolean",
            "title": "Interrupt",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "role",
          "type",
          "position_x",
          "position_y"
        ],
        "title": "MemberCreate"
      },
      "MemberOut": {
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,64}$",
            "title": "Name"
          },
          "backstory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backstory"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "owner_of": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Of"
          },
          "position_x": {
            "type": "number",
            "title": "Position X"
          },
          "position_y": {
            "type": "number",
            "title": "Position Y"
          },
          "source": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "zhipuai"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "glm-4-0520"
          },
          "openai_api_key": {
            "type": "string",
            "title": "Openai Api Key",
            "default": "fe1f2097b7284bd4baa1284be8d54aea.6VOvX4efbye8M6m0"
          },
          "openai_api_base": {
            "type": "string",
            "title": "Openai Api Base",
            "default": "https://open.bigmodel.cn/api/paas/v4"
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "default": 0.1
          },
          "interrupt": {
            "type": "boolean",
            "title": "Interrupt",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "belongs_to": {
            "type": "integer",
            "title": "Belongs To"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/Skill"
            },
            "type": "array",
            "title": "Skills"
          },
          "uploads": {
            "items": {
              "$ref": "#/components/schemas/Upload"
            },
            "type": "array",
            "title": "Uploads"
          }
        },
        "type": "object",
        "required": [
          "name",
          "role",
          "type",
          "owner_of",
          "position_x",
          "position_y",
          "id",
          "belongs_to",
          "skills",
          "uploads"
        ],
        "title": "MemberOut"
      },
      "MemberUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-zA-Z0-9_-]{1,64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "backstory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backstory"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "owner_of": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Of"
          },
          "position_x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position X"
          },
          "position_y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Y"
          },
          "source": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "openai_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openai Api Key"
          },
          "openai_api_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openai Api Base"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "interrupt": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt"
          },
          "belongs_to": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Belongs To"
          },
          "skills": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Skill"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skills"
          },
          "uploads": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Upload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploads"
          }
        },
        "type": "object",
        "title": "MemberUpdate"
      },
      "MembersOut": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MemberOut"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "MembersOut"
      },
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
...
      "ModelProviderCreate": {
        "properties": {
          "provider_name": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,64}$",
            "title": "Provider Name",
            "unique": true
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "provider_name",
          "description"
        ],
        "title": "ModelProviderCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}